requirements.txt
-i https://pypi.org/simple
altair==5.3.0; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
blinker==1.8.2; python_version >= '3.8'
cachetools==5.4.0; python_version >= '3.7'
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
distro==1.9.0; python_version >= '3.6'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.43; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
jinja2==3.1.4; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
numpy==2.0.1; python_version >= '3.9'
openai==1.36.1; python_full_version >= '3.7.1'
packaging==24.1; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
pillow==10.4.0; python_version >= '3.8'
protobuf==5.27.2; python_version >= '3.8'
pyarrow==17.0.0; python_version >= '3.8'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pydeck==0.9.1; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
referencing==0.35.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rich==13.7.1; python_full_version >= '3.7.0'
rpds-py==0.19.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
streamlit==1.36.0; python_version >= '3.8' and python_full_version != '3.9.7'
tenacity==8.5.0; python_version >= '3.8'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.12.1; python_version >= '3.7'
tornado==6.4.1; python_version >= '3.8'
tqdm==4.66.4; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
urllib3==2.2.2; python_version >= '3.8'
